Split items into modules
