+++ src/lib.rs
@@ -1,13 +1,18 @@
 #![doc = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/README.md"))]
 
 use std::{
-    cmp,
     fmt::{self, Display, Formatter},
     fs::File,
-    io::{BufRead, BufReader, Cursor, Read, Result, Seek, SeekFrom},
+    io::{BufRead, BufReader, Cursor, Read, Result, Seek},
     path::Path,
     str,
 };
+
+#[macro_use]
+mod macros;
+
+mod binary;
+mod signatures;
 
 /// A file format.
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -1354,137 +1359,6 @@ impl FileFormat {
             })
             .unwrap_or_default())
     }
-
-    /// Determines `FileFormat` from a Compound File Binary reader.
-    #[cfg(feature = "cfb")]
-    fn from_cfb<R: Read + Seek>(reader: R) -> Result<Self> {
-        let file = cfb::CompoundFile::open(reader)?;
-        Ok(match file.root_entry().clsid().to_string().as_str() {
-            "00020810-0000-0000-c000-000000000046" => Self::MicrosoftExcelSpreadsheet,
-            "00020820-0000-0000-c000-000000000046" => Self::MicrosoftExcelSpreadsheet,
-            "64818d10-4f9b-11cf-86ea-00aa00b929e8" => Self::MicrosoftPowerpointPresentation,
-            "74b78f3a-c8c8-11d1-be11-00c04fb6faf1" => Self::MicrosoftProjectPlan,
-            "00021201-0000-0000-00c0-000000000046" => Self::MicrosoftPublisherDocument,
-            "000c1084-0000-0000-c000-000000000046" => Self::MicrosoftSoftwareInstaller,
-            "00021a14-0000-0000-c000-000000000046" => Self::MicrosoftVisioDrawing,
-            "00020906-0000-0000-c000-000000000046" => Self::MicrosoftWordDocument,
-            _ => Self::CompoundFileBinary,
-        })
-    }
-
-    /// Determines `FileFormat` from a Matroska Video reader.
-    fn from_mkv<R: Read + Seek>(mut reader: R) -> Result<Self> {
-        let length = reader.seek(SeekFrom::End(0))?;
-        reader.rewind()?;
-        let mut buf = vec![0; cmp::min(4096, length as usize)];
-        reader.read_exact(&mut buf)?;
-        if buf
-            .windows(2)
-            .position(|bytes| bytes == b"\x42\x82")
-            .map(|pos| &buf[pos + 3..pos + 7])
-            .filter(|bytes| bytes == b"webm")
-            .is_some()
-        {
-            Ok(Self::Webm)
-        } else {
-            Ok(Self::MatroskaVideo)
-        }
-    }
-
-    /// Determines `FileFormat` from a MS-DOS Executable reader.
-    fn from_ms_dos_executable<R: Read + Seek>(mut reader: R) -> Result<Self> {
-        reader.seek(SeekFrom::Start(0x3C))?;
-        let mut address = [0; 4];
-        reader.read_exact(&mut address)?;
-        reader.seek(SeekFrom::Start(u32::from_le_bytes(address) as u64))?;
-        let mut signature = [0; 4];
-        reader.read_exact(&mut signature)?;
-        if &signature == b"PE\x00\x00" {
-            reader.seek(SeekFrom::Current(0x12))?;
-            let mut characteristics = [0; 2];
-            reader.read_exact(&mut characteristics)?;
-            return Ok(if u16::from_le_bytes(characteristics) & 0x2000 == 0x2000 {
-                Self::DynamicLinkLibrary
-            } else {
-                Self::PortableExecutable
-            });
-        }
-        Ok(Self::MsDosExecutable)
-    }
-
-    /// Determines `FileFormat` from a Portable Document Format reader.
-    fn from_pdf<R: Read + Seek>(mut reader: R) -> Result<Self> {
-        let length = reader.seek(SeekFrom::End(0))?;
-        reader.rewind()?;
-        let mut buf = vec![0; cmp::min(10 * 1024 * 1024, length as usize)];
-        reader.read_exact(&mut buf)?;
-        Ok(if buf.windows(13).any(|bytes| bytes == b"AIPrivateData") {
-            Self::AdobeIllustratorArtwork
-        } else {
-            Self::PortableDocumentFormat
-        })
-    }
-
-    /// Determines `FileFormat` from a ZIP reader.
-    #[cfg(feature = "zip")]
-    fn from_zip<R: Read + Seek>(reader: R) -> Result<Self> {
-        let mut archive = zip::ZipArchive::new(reader)?;
-        let mut format = Self::Zip;
-        for index in 0..archive.len() {
-            let mut file = archive.by_index(index)?;
-            match file.name() {
-                "AndroidManifest.xml" => return Ok(Self::AndroidPackage),
-                "AppManifest.xaml" => return Ok(Self::Xap),
-                "AppxManifest.xml" => return Ok(Self::WindowsAppPackage),
-                "META-INF/MANIFEST.MF" => format = Self::JavaArchive,
-                "META-INF/application.xml" => return Ok(Self::EnterpriseApplicationArchive),
-                "META-INF/mozilla.rsa" => return Ok(Self::Xpinstall),
-                "WEB-INF/web.xml" => return Ok(Self::WebApplicationArchive),
-                "doc.kml" => return Ok(Self::KeyholeMarkupLanguageZipped),
-                "extension.vsixmanifest" => return Ok(Self::MicrosoftVisualStudioExtension),
-                "mimetype" => {
-                    let mut content = String::new();
-                    file.read_to_string(&mut content)?;
-                    match content.trim() {
-                        "application/epub+zip" => return Ok(Self::ElectronicPublication),
-                        "application/vnd.oasis.opendocument.graphics" => {
-                            return Ok(Self::OpenDocumentGraphics)
-                        }
-                        "application/vnd.oasis.opendocument.presentation" => {
-                            return Ok(Self::OpenDocumentPresentation);
-                        }
-                        "application/vnd.oasis.opendocument.spreadsheet" => {
-                            return Ok(Self::OpenDocumentSpreadsheet);
-                        }
-                        "application/vnd.oasis.opendocument.text" => {
-                            return Ok(Self::OpenDocumentText);
-                        }
-                        _ => {}
-                    }
-                }
-                _ => {
-                    if file.name().starts_with("circuitdiagram/") {
-                        return Ok(Self::CircuitDiagramDocument);
-                    } else if file.name().starts_with("dwf/") {
-                        return Ok(Self::DesignWebFormatXps);
-                    } else if file.name().starts_with("word/") {
-                        return Ok(Self::OfficeOpenXmlDocument);
-                    } else if file.name().starts_with("visio/") {
-                        return Ok(Self::OfficeOpenXmlDrawing);
-                    } else if file.name().starts_with("ppt/") {
-                        return Ok(Self::OfficeOpenXmlPresentation);
-                    } else if file.name().starts_with("xl/") {
-                        return Ok(Self::OfficeOpenXmlSpreadsheet);
-                    } else if file.name().starts_with("3D/") && file.name().ends_with(".model") {
-                        return Ok(Self::ThreeDimensionalManufacturingFormat);
-                    } else if file.name().starts_with("Payload/") && file.name().contains(".app/") {
-                        return Ok(Self::IosAppStorePackage);
-                    }
-                }
-            }
-        }
-        Ok(format)
-    }
 }
 
 impl Default for FileFormat {
@@ -1506,631 +1380,6 @@ impl From<&[u8]> for FileFormat {
     fn from(value: &[u8]) -> Self {
         Self::from_reader(Cursor::new(value)).unwrap_or_default()
     }
-}
-
-/// Generates [`FileFormat::from_signature`] function.
-macro_rules! signatures {
-    {
-        $(
-            format = $format:ident
-            $(value = $($value:literal $(offset = $offset:literal)?),+)+
-        )*
-    } => {
-        impl FileFormat {
-            /// Determines `FileFormat` by checking its signature.
-            fn from_signature(bytes: &[u8]) -> Option<Self> {
-                $(
-                    if $($(bytes.len() >= $($offset +)? $value.len()
-                        && &bytes[$($offset)?..$($offset +)? $value.len()] == $value)&&*)||* {
-                        return Some(Self::$format);
-                    }
-                )*
-                None
-            }
-        }
-    };
-}
-
-signatures! {
-    // 39-byte signatures
-    format = VirtualboxVirtualDiskImage
-    value = b"<<< Oracle VM VirtualBox Disk Image >>>"
-
-    // 32-byte signatures
-    format = Sketchup
-    value =
-        b"\xFF\xFE\xFF\x0E\x53\x00\x6B\x00\x65\x00\x74\x00\x63\x00\x68\x00",
-        b"\x55\x00\x70\x00\x20\x00\x4D\x00\x6F\x00\x64\x00\x65\x00\x6C\x00" offset = 16
-
-    // 29-byte signatures
-    format = FlexibleImageTransportSystem
-    value =
-        b"\x49\x4D\x50\x4C\x45\x20\x20\x3D\x20\x20\x20\x20\x20\x20\x20",
-        b"\x20\x20\x20\x20\x20\x20\x20\x20\x20\x20\x20\x20\x20\x54" offset = 15
-
-    // 22-byte signatures
-    format = CreativeVoice
-    value = b"Creative Voice File\x1A\x1A\x00"
-
-    // 21-byte signatures
-    format = DebianBinaryPackage
-    value = b"!<arch>\ndebian-binary"
-
-    // 20-byte signatures
-    format = WindowsShortcut
-    value = b"\x4C\x00\x00\x00\x01\x14\x02\x00\x00\x00\x00\x00\xC0\x00\x00\x00\x00\x00\x00\x46"
-
-    // 16-byte signatures
-    format = AdobeIndesignDocument
-    value = b"\x06\x06\xED\xF5\xD8\x1D\x46\xE5\xBD\x31\xEF\xE7\xFE\x74\xB7\x1D"
-
-    format = FastTracker2ExtendedModule
-    value = b"Extended Module:"
-
-    format = MacosAlias
-    value = b"\x62\x6F\x6F\x6B\x00\x00\x00\x00\x6D\x61\x72\x6B\x00\x00\x00\x00"
-
-    format = Sqlite3
-    value = b"\x53\x51\x4C\x69\x74\x65\x20\x66\x6F\x72\x6D\x61\x74\x20\x33\x00"
-
-    format = WindowsMediaVideo
-    value = b"\x30\x26\xB2\x75\x8E\x66\xCF\x11\xA6\xD9\x00\xAA\x00\x62\xCE\x6C"
-
-    // 15-byte signatures
-    format = FujifilmRaw
-    value = b"FUJIFILMCCD-RAW"
-
-    // 14-byte signatures
-    format = MaterialExchangeFormat
-    value = b"\x06\x0E\x2B\x34\x02\x05\x01\x01\x0D\x01\x02\x01\x01\x02"
-
-    // 12-byte signatures
-    format = AnimatedPortableNetworkGraphics
-    value = b"\x89\x50\x4E\x47\x0D\x0A\x1A\x0A", b"acTL" offset = 0x25
-
-    format = Djvu
-    value = b"AT&TFORM", b"DJVM" offset = 12
-    value = b"AT&TFORM", b"DJVU" offset = 12
-    value = b"AT&TFORM", b"DJVI" offset = 12
-    value = b"AT&TFORM", b"THUM" offset = 12
-
-    format = JpegXl
-    value = b"\x00\x00\x00\x0C\x4A\x58\x4C\x20\x0D\x0A\x87\x0A"
-    value = b"\xFF\x0A"
-
-    format = KhronosTexture
-    value = b"\xAB\x4B\x54\x58\x20\x31\x31\xBB\x0D\x0A\x1A\x0A"
-
-    format = KhronosTexture2
-    value = b"\xAB\x4B\x54\x58\x20\x32\x30\xBB\x0D\x0A\x1A\x0A"
-
-    format = MatroskaVideo
-    value = b"\x1A\x45\xDF\xA3"
-
-    format = OggOpus
-    value = b"OggS", b"OpusHead" offset = 28
-
-    format = PanasonicRaw
-    value = b"\x49\x49\x55\x00\x18\x00\x00\x00\x88\xE7\x74\xD8"
-
-    // 11-byte signatures
-    format = OggSpeex
-    value = b"OggS", b"Speex  " offset = 28
-
-    format = OggTheora
-    value = b"OggS", b"\x80\x74\x68\x65\x6F\x72\x61" offset = 28
-
-    format = OggVorbis
-    value = b"OggS", b"\x01\x76\x6F\x72\x62\x69\x73" offset = 28
-
-    format = RadianceHdr
-    value = b"\x23\x3F\x52\x41\x44\x49\x41\x4E\x43\x45\x0A"
-
-    // 10-byte signatures
-    format = OggMedia
-    value = b"OggS", b"\x01\x76\x69\x64\x65\x6F" offset = 28
-
-    format = Snappy
-    value = b"\xFF\x06\x00\x00\x73\x4E\x61\x50\x70\x59"
-
-    // 9-byte signatures
-    format = GameBoyColorRom
-    value = b"\xCE\xED\x66\x66\xCC\x0D\x00\x0B" offset = 0x104, b"\x80" offset = 0x143
-    value = b"\xCE\xED\x66\x66\xCC\x0D\x00\x0B" offset = 0x104, b"\xC0" offset = 0x143
-
-    format = Lzop
-    value = b"\x89\x4C\x5A\x4F\x00\x0D\x0A\x1A\x0A"
-
-    format = MicrosoftVirtualHardDisk
-    value = b"connectix"
-
-    format = OggFlac
-    value = b"OggS", b"\x7F\x46\x4C\x41\x43" offset = 28
-
-    format = OlympusRawFormat
-    value = b"\x49\x49\x52\x4F\x08\x00\x00\x00\x18"
-
-    // 8-byte signatures
-    format = Ani
-    value = b"RIFF", b"ACON" offset = 8
-
-    format = AudioInterchangeFileFormat
-    value = b"FORM", b"AIFF" offset = 8
-    value = b"FORM", b"AIFC" offset = 8
-
-    format = AudioVideoInterleave
-    value = b"RIFF", b"\x41\x56\x49\x20" offset = 8
-
-    format = Av1ImageFileFormat
-    value = b"ftypavif" offset = 4
-
-    format = Av1ImageFileFormatSequence
-    value = b"ftypavis" offset = 4
-
-    format = CompoundFileBinary
-    value = b"\xD0\xCF\x11\xE0\xA1\xB1\x1A\xE1"
-
-    format = DalvikExecutable
-    value = b"\x64\x65\x78\x0A\x30\x33\x35\x00"
-
-    format = ExperimentalComputingFacility
-    value = b"gimp xcf"
-
-    format = GameBoyAdvanceRom
-    value = b"\x24\xFF\xAE\x51\x69\x9A\xA2\x21" offset = 4
-
-    format = GameBoyRom
-    value = b"\xCE\xED\x66\x66\xCC\x0D\x00\x0B" offset = 0x104
-
-    format = HighEfficiencyImageCoding
-    value = b"ftypheic" offset = 4
-    value = b"ftypheix" offset = 4
-
-    format = HighEfficiencyImageCodingSequence
-    value = b"ftyphevc" offset = 4
-    value = b"ftyphevx" offset = 4
-
-    format = HighEfficiencyImageFileFormat
-    value = b"ftypmif1" offset = 4
-
-    format = HighEfficiencyImageFileFormatSequence
-    value = b"ftypmsf1" offset = 4
-
-    format = Jpeg2000Part1
-    value = b"ftypjp2 " offset = 16
-
-    format = Jpeg2000Part2
-    value = b"ftypjpx " offset = 16
-
-    format = Jpeg2000Part3
-    value = b"ftypmjp2" offset = 16
-
-    format = Jpeg2000Part6
-    value = b"ftypjpm " offset = 16
-
-    format = MicrosoftVirtualHardDisk2
-    value = b"vhdxfile"
-
-    format = Mobipocket
-    value = b"BOOKMOBI" offset = 60
-
-    format = Mpeg4Part14Video
-    value = b"ftypavc1" offset = 4
-    value = b"ftypdash" offset = 4
-    value = b"ftypiso2" offset = 4
-    value = b"ftypiso3" offset = 4
-    value = b"ftypiso4" offset = 4
-    value = b"ftypiso5" offset = 4
-    value = b"ftypiso6" offset = 4
-    value = b"ftypisom" offset = 4
-    value = b"ftypmmp4" offset = 4
-    value = b"ftypmp41" offset = 4
-    value = b"ftypmp42" offset = 4
-    value = b"ftypmp4v" offset = 4
-    value = b"ftypmp71" offset = 4
-    value = b"ftypMSNV" offset = 4
-    value = b"ftypNDAS" offset = 4
-    value = b"ftypNDSC" offset = 4
-    value = b"ftypNDSH" offset = 4
-    value = b"ftypNDSM" offset = 4
-    value = b"ftypNDSP" offset = 4
-    value = b"ftypNDSS" offset = 4
-    value = b"ftypNDXC" offset = 4
-    value = b"ftypNDXH" offset = 4
-    value = b"ftypNDXM" offset = 4
-    value = b"ftypNDXP" offset = 4
-
-    format = NikonElectronicFile
-    value = b"\x4D\x4D\x00\x2A", b"\x1C\x00\xFE\x00" offset = 8
-    value = b"\x4D\x4D\x00\x2A", b"\x1F\x00\x0B\x00" offset = 8
-    value = b"\x49\x49\x2A\x00", b"\x1C\x00\xFE\x00" offset = 8
-    value = b"\x49\x49\x2A\x00", b"\x1F\x00\x0B\x00" offset = 8
-
-    format = Nintendo64Rom
-    value = b"\x80\x37\x12\x40\x00\x00\x00\x0F"
-    value = b"\x37\x80\x40\x12\x00\x00\x0F\x00"
-    value = b"\x12\x40\x80\x37\x00\x0F\x00\x00"
-    value = b"\x40\x12\x37\x80\x0F\x00\x00\x00"
-
-    format = NintendoDsRom
-    value = b"\x24\xFF\xAE\x51\x69\x9A\xA2\x21" offset = 0xC0
-    value = b"\xC8\x60\x4F\xE2\x01\x70\x8F\xE2" offset = 0xC0
-
-    format = PortableNetworkGraphics
-    value = b"\x89\x50\x4E\x47\x0D\x0A\x1A\x0A"
-
-    format = QualcommPureVoice
-    value = b"RIFF", b"QLCM" offset = 8
-
-    format = RoshalArchive
-    value = b"\x52\x61\x72\x21\x1A\x07\x01\x00"
-    value = b"\x52\x61\x72\x21\x1A\x07\x00"
-
-    format = TapeArchive
-    value = b"\x75\x73\x74\x61\x72\x00\x30\x30" offset = 257
-    value = b"\x75\x73\x74\x61\x72\x20\x20\x00" offset = 257
-
-    format = WaveformAudio
-    value = b"RIFF", b"WAVE" offset = 8
-
-    format = Webp
-    value = b"RIFF", b"WEBP" offset = 8
-
-    // 7-byte signatures
-    format = AdobeFlashPlayerAudio
-    value = b"ftypF4A" offset = 4
-
-    format = AdobeFlashPlayerAudiobook
-    value = b"ftypF4B" offset = 4
-
-    format = AdobeFlashPlayerProtectedVideo
-    value = b"ftypF4P" offset = 4
-
-    format = AdobeFlashPlayerVideo
-    value = b"ftypF4V" offset = 4
-
-    format = AppleItunesAudio
-    value = b"ftypM4A" offset = 4
-
-    format = AppleItunesAudiobook
-    value = b"ftypM4B" offset = 4
-
-    format = AppleItunesProtectedAudio
-    value = b"ftypM4P" offset = 4
-
-    format = AppleItunesVideo
-    value = b"ftypM4V" offset = 4
-
-    format = Blender
-    value = b"BLENDER"
-
-    format = CanonRaw3
-    value = b"ftypcrx" offset = 4
-
-    format = MetaInformationEncapsulation
-    value = b"\x7E\x10\x04", b"\x30\x4D\x49\x45" offset = 4
-    value = b"\x7E\x18\x04", b"\x30\x4D\x49\x45" offset = 4
-
-    format = SonyMovie
-    value = b"ftypmqt" offset = 4
-
-    format = ThirdGenerationPartnershipProject
-    value = b"ftyp3gp" offset = 4
-
-    format = ThirdGenerationPartnershipProject2
-    value = b"ftyp3g2" offset = 4
-
-    format = UnixArchiver
-    value = b"!<arch>"
-
-    // 6-byte signatures
-    format = ApacheArrowColumnar
-    value = b"ARROW1"
-
-    format = AppleQuicktime
-    value = b"ftypqt" offset = 4
-    value = b"moov" offset = 4
-    value = b"mdat" offset = 4
-    value = b"wide" offset = 4
-    value = b"skip" offset = 4
-    value = b"free" offset = 4
-
-    format = CanonRaw2
-    value = b"\x4D\x4D\x00\x2A", b"CR" offset = 8
-    value = b"\x49\x49\x2A\x00", b"CR" offset = 8
-
-    format = GraphicsInterchangeFormat
-    value = b"GIF87a"
-    value = b"GIF89a"
-
-    format = SevenZip
-    value = b"\x37\x7A\xBC\xAF\x27\x1C"
-    value = b"\x37\x7A\xBC\xAF\x27\x1C"
-
-    format = Xz
-    value = b"\xFD\x37\x7A\x58\x5A\x00"
-
-    // 5-byte signatures
-    format = AdaptiveMultiRate
-    value = b"#!AMR"
-
-    format = EmbeddedOpentype
-    value = b"\x00\x00\x01" offset = 8, b"\x4C\x50" offset = 34
-    value = b"\x01\x00\x02" offset = 8, b"\x4C\x50" offset = 34
-    value = b"\x02\x00\x02" offset = 8, b"\x4C\x50" offset = 34
-
-    format = Iso9660
-    value = b"CD001" offset = 0x8001
-    value = b"CD001" offset = 0x8801
-    value = b"CD001" offset = 0x9001
-
-    format = Lha
-    value = b"-lh0-" offset = 2
-    value = b"-lh1-" offset = 2
-    value = b"-lh2-" offset = 2
-    value = b"-lh3-" offset = 2
-    value = b"-lh4-" offset = 2
-    value = b"-lh5-" offset = 2
-    value = b"-lh6-" offset = 2
-    value = b"-lh7-" offset = 2
-    value = b"-lzs-" offset = 2
-    value = b"-lz4-" offset = 2
-    value = b"-lz5-" offset = 2
-    value = b"-lhd-" offset = 2
-
-    format = Opentype
-    value = b"\x4F\x54\x54\x4F\x00"
-
-    format = PortableDocumentFormat
-    value = b"%PDF-"
-
-    format = Truetype
-    value = b"\x00\x01\x00\x00\x00"
-
-    // 4-byte signatures
-    format = AdobePhotoshopDocument
-    value = b"8BPS"
-
-    format = Alz
-    value = b"\x41\x4C\x5A\x01"
-
-    format = AndroidBinaryXml
-    value = b"\x03\x00\x08\x00"
-
-    format = AndroidCompiledResources
-    value = b"\x02\x00\x0C\x00"
-
-    format = AppleIconImage
-    value = b"icns"
-
-    format = Au
-    value = b".snd"
-
-    format = BetterPortableGraphics
-    value = b"\x42\x50\x47\xFB"
-
-    format = Cabinet
-    value = b"MSCF"
-    value = b"ISc("
-
-    format = Cineon
-    value = b"\x80\x2A\x5F\xD7"
-
-    format = Cur
-    value = b"\x00\x00\x02\x00"
-
-    format = DigitalImagingAndCommunicationsInMedicine
-    value = b"\x44\x49\x43\x4D" offset = 128
-
-    format = DigitalPictureExchange
-    value = b"SDPX"
-    value = b"XPDS"
-
-    format = ExecutableAndLinkableFormat
-    value = b"\x7F\x45\x4C\x46"
-
-    format = ExtensibleArchive
-    value = b"xar!"
-
-    format = FlashVideo
-    value = b"\x46\x4C\x56\x01"
-
-    format = FreeLosslessAudioCodec
-    value = b"fLaC"
-
-    format = FreeLosslessImageFormat
-    value = b"FLIF"
-
-    format = GlTransmissionFormatBinary
-    value = b"glTF"
-
-    format = GoogleChromeExtension
-    value = b"Cr24"
-
-    format = Ico
-    value = b"\x00\x00\x01\x00"
-
-    format = ImpulseTrackerModule
-    value = b"IMPM"
-
-    format = JavaClass
-    value = b"\xCA\xFE\xBA\xBE"
-
-    format = JavaKeystore
-    value = b"\xFE\xED\xFE\xED"
-
-    format = LempelZivFiniteStateEntropy
-    value = b"bvx-"
-    value = b"bvx1"
-    value = b"bvx2"
-    value = b"bvxn"
-
-    format = LongRangeZip
-    value = b"LRZI"
-
-    format = LuaBytecode
-    value = b"\x1B\x4C\x75\x61"
-
-    format = Lz4
-    value = b"\x04\x22\x4D\x18"
-
-    format = Lzip
-    value = b"LZIP"
-
-    format = MicrosoftCompiledHtmlHelp
-    value = b"ITSF"
-
-    format = MicrosoftDirectDrawSurface
-    value = b"DDS "
-
-    format = MonkeysAudio
-    value = b"MAC "
-
-    format = Mpeg1Video
-    value = b"\x00\x00\x01\xBA"
-    value = b"\x00\x00\x01\xB3"
-
-    format = Musepack
-    value = b"MPCK"
-    value = b"MP+"
-
-    format = MusicalInstrumentDigitalInterface
-    value = b"MThd"
-
-    format = NintendoEntertainmentSystemRom
-    value = b"\x4E\x45\x53\x1A"
-
-    format = OggMultiplexedMedia
-    value = b"OggS"
-
-    format = Openexr
-    value = b"\x76\x2F\x31\x01"
-
-    format = OptimizedDalvikExecutable
-    value = b"dey\n"
-
-    format = PcapDump
-    value = b"\xA1\xB2\xC3\xD4"
-    value = b"\xD4\xC3\xB2\xA1"
-
-    format = PcapNextGenerationDump
-    value = b"\x0A\x0D\x0D\x0A"
-
-    format = RedHatPackageManager
-    value = b"\xED\xAB\xEE\xDB"
-
-    format = Screamtracker3Module
-    value = b"SCRM" offset = 44
-
-    format = Shapefile
-    value = b"\x00\x00\x27\x0A"
-
-    format = SonyDsdStreamFile
-    value = b"DSD "
-
-    format = TagImageFileFormat
-    value = b"\x4D\x4D\x00\x2A"
-    value = b"\x49\x49\x2A\x00"
-    value = b"\x4D\x4D\x00\x2B"
-    value = b"\x49\x49\x2B\x00"
-
-    format = Tasty
-    value = b"\x5C\xA1\xAB\x1F"
-
-    format = Wavpack
-    value = b"wvpk"
-
-    format = WebassemblyBinary
-    value = b"\x00\x61\x73\x6D"
-
-    format = WebOpenFontFormat
-    value = b"wOFF"
-
-    format = WebOpenFontFormat2
-    value = b"wOF2"
-
-    format = WindowsMetafile
-    value = b"\xD7\xCD\xC6\x9A"
-    value = b"\x02\x00\x09\x00"
-    value = b"\x01\x00\x09\x00"
-
-    format = Zip
-    value = b"\x50\x4B\x03\x04"
-
-    format = Zstandard
-    value = b"\x28\xB5\x2F\xFD"
-
-    // 3-byte signatures
-    format = Bzip2
-    value = b"BZh"
-
-    format = JointPhotographicExpertsGroup
-    value = b"\xFF\xD8\xFF"
-
-    format = JpegExtendedRange
-    value = b"\x49\x49\xBC"
-
-    format = Mpeg12AudioLayer3
-    value = b"ID3"
-    value = b"\xFF\xE3"
-    value = b"\xFF\xF3"
-    value = b"\xFF\xFB"
-
-    format = Seqbox
-    value = b"SBx"
-
-    format = SmallWebFormat
-    value = b"\x43\x57\x53"
-    value = b"\x46\x57\x53"
-
-    format = Zoo
-    value = b"ZOO"
-
-    // 2-byte signatures
-    format = AdvancedAudioCoding
-    value = b"\xFF\xF1"
-    value = b"\xFF\xF9"
-
-    format = AppleDiskImage
-    value = b"\x78\x01"
-
-    format = ArchivedByRobertJung
-    value = b"\x60\xEA"
-
-    format = AudioCodec3
-    value = b"\x0B\x77"
-
-    format = Cpio
-    value = b"\xC7\x71"
-    value = b"\x71\xC7"
-
-    format = Gzip
-    value = b"\x1F\x8B"
-
-    format = Mpeg1AudioLayer1
-    value = b"\xFF\xFE"
-    value = b"\xFF\xF6"
-    value = b"\xFF\xFF"
-
-    format = Mpeg1AudioLayer2
-    value = b"\xFF\xF4"
-    value = b"\xFF\xF5"
-    value = b"\xFF\xFC"
-    value = b"\xFF\xFD"
-
-    format = Mpeg2TransportStream
-    value = b"\x47", b"\x47" offset = 188
-    value = b"\x47" offset = 4, b"\x47" offset = 196
-
-    format = MsDosExecutable
-    value = b"MZ"
-
-    format = UnixCompress
-    value = b"\x1F\xA0"
-    value = b"\x1F\x9D"
-
-    format = WindowsBitmap
-    value = b"BM"
 }
 
 /// Kind of [`FileFormat`].
