+++ src/macros.rs
@@ -1,0 +1,24 @@
+//! Convenient macros.
+
+/// Generates [`FileFormat::from_signature`] function.
+macro_rules! signatures {
+    {
+        $(
+            format = $format:ident
+            $(value = $($value:literal $(offset = $offset:literal)?),+)+
+        )*
+    } => {
+        impl crate::FileFormat {
+            /// Determines `FileFormat` by checking its signature.
+            pub(crate) fn from_signature(bytes: &[u8]) -> Option<Self> {
+                $(
+                    if $($(bytes.len() >= $($offset +)? $value.len()
+                        && &bytes[$($offset)?..$($offset +)? $value.len()] == $value)&&*)||* {
+                        return Some(Self::$format);
+                    }
+                )*
+                None
+            }
+        }
+    };
+}
